avoid non-Racket-syntax causing reader errors
The old reader used a Racket read function and parsed the resulting
syntax tree; this meant that any non-Racket-syntax would cause a read
error.

The new, hand-written reader only reads as much as it needs using
Racket's readers, so avoids this problem (on a reader error it
assumes that there are no more directives to read and can finish its
job).

One quirk is that, while `#lang`s are expected to consume a whole file
with `racket <file>`, they need not do so with, say, `racket -f <file>`,
because each expression is evaluated (including any not parsed as part
of the `#reader` protocol to which `#lang` effectively expands). This
means that my `#lang`s reader needs to consume the whole port---closing
it won't do---to avoid extraneous expressions appearing and gumming up
the semantics of `racket <file>`.